+++ mainapp/urls.py
@@ -1,9 +1,7 @@
 from django.urls import path
 from django.conf.urls import url
 from django.contrib.auth import views as auth_views
-
 from . import views
-
 
 urlpatterns = [
     path('', views.HomePageView.as_view(), name='home'),
